Fix issues and add docs

+++ src/app/redux/effects/app-stations.effects.ts
@@ -2,16 +2,18 @@ import { inject, Injectable } from '@angular/core';
 import { Actions, createEffect, ofType } from '@ngrx/effects';
 import { concatLatestFrom } from '@ngrx/operators';
 import { Store } from '@ngrx/store';
-import { catchError, EMPTY, endWith, exhaustMap, map, of, startWith } from 'rxjs';
+import { catchError, endWith, exhaustMap, map, of, startWith } from 'rxjs';
 
 import { StationsService } from '../../admin/services/stations.service';
 import { Station } from '../../core/models/station/station.model';
 import { MessagesService } from '../../core/services/messages.service';
 import { AppConfigActions } from '../actions/app-config.actions';
+import { AppOrdersActions } from '../actions/app-orders.actions';
 import { AppRoutesActions } from '../actions/app-routes.actions';
+import { AppSchedulesActions } from '../actions/app-schedule.actions';
 import { AppStationsActions } from '../actions/app-station.actions';
 import { AppTripActions } from '../actions/app-trip.actions';
-import { selectStations } from '../selectors/app-stations.selector';
+import { selectIsStationInActiveRide, selectStations } from '../selectors/app-stations.selector';
 
 @Injectable()
 export class AppStationsEffects {
@@ -28,14 +30,8 @@ export class AppStationsEffects {
 
     loadStations$ = createEffect(() =>
         this.actions$.pipe(
-            ofType(
-                AppStationsActions.loadStations,
-                AppStationsActions.deleteStationSuccess,
-                AppRoutesActions.loadRoutesSuccess,
-                AppTripActions.loadTripInfoSuccess
-            ),
+            ofType(AppStationsActions.loadStations, AppStationsActions.deleteStationSuccess),
             exhaustMap(() => {
-                console.log('\x1b[31m%s\x1b[0m', 'liflud');
                 return this.stationsService.getStations().pipe(
                     map((stations: Station[]) => {
                         this.form.reset();
@@ -50,11 +46,17 @@ export class AppStationsEffects {
     );
     lazyLoadStations$ = createEffect(() =>
         this.actions$.pipe(
-            ofType(AppStationsActions.lazyLoadStations, AppRoutesActions.loadRoutesSuccess),
+            ofType(
+                AppStationsActions.lazyLoadStations,
+                AppRoutesActions.loadRoutesSuccess,
+                AppSchedulesActions.loadSchedulesSuccess,
+                AppOrdersActions.loadOrdersSuccess,
+                AppTripActions.loadTripInfoSuccess
+            ),
             concatLatestFrom(() => this.store.select(selectStations)),
             exhaustMap(([, stationsOld]) => {
                 if (stationsOld && stationsOld.length > 0) {
-                    return EMPTY;
+                    return of(AppStationsActions.stationsLoadNotRequired());
                 }
                 return this.stationsService.getStations().pipe(
                     map((stations: Station[]) => {
@@ -74,10 +76,10 @@ export class AppStationsEffects {
             ofType(AppStationsActions.initSaveNewStation),
             exhaustMap((action) => {
                 return this.stationsService.postStation(action.station).pipe(
-                    map(() => {
+                    map(({ id }) => {
                         this.form.reset();
                         this.messagesService.sendSuccess('MESSAGES.STATIONS.SAVE_SUCCESS');
-                        return AppStationsActions.newStationSavedSuccess({ station: action.station });
+                        return AppStationsActions.newStationSavedSuccess({ station: { ...action.station, id } });
                     }),
                     catchError((error) => {
                         return of(AppStationsActions.newStationSavedFailure({ error }));
@@ -108,10 +110,19 @@ export class AppStationsEffects {
         )
     );
 
-    deleteStation$ = createEffect(() =>
+    deleteOrder$ = createEffect(() =>
         this.actions$.pipe(
             ofType(AppStationsActions.initDeleteStation),
-            exhaustMap((action) => {
+            concatLatestFrom((action) => this.store.select(selectIsStationInActiveRide(action.id))),
+            exhaustMap(([action, isStationInActiveRide]) => {
+                if (isStationInActiveRide) {
+                    this.messagesService.sendError('MESSAGES.STATIONS.DELETE_STATION_IN_ACTIVE_RIDE');
+                    return of(
+                        AppOrdersActions.cancelOrderFailure({
+                            error: 'MESSAGES.STATIONS.DELETE_STATION_IN_ACTIVE_RIDE',
+                        })
+                    );
+                }
                 return this.stationsService.deleteStation(action.id).pipe(
                     map(() => {
                         this.messagesService.sendSuccess('MESSAGES.STATIONS.DELETE_SUCCESS');
